requirements.txt
httpx>=0.28.1
tenacity>=9.0.0
loguru>=0.7.0
pydantic>=2.11.0
typer[all]>=0.12.5
tqdm>=4.66.0
numpy>=1.26.0
Pillow>=10.0.0
pymupdf>=1.24.0
sentence-transformers>=2.7.0
open-clip-torch>=2.24.0
qdrant-client>=1.9.0
arxiv>=2.1.0
neo4j>=5.23.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
groq>=0.31.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.3.0
langgraph>=0.6.0

